ai-telegram-bot/requirements.txt
python-telegram-bot==20.7
openai==1.6.0
python-dotenv==1.0.0
anthropic==0.8.0